+++ KPI-ejb/src/java/cn/hanbell/kpi/mail/BscGroupShipmentMailBean.java
@@ -6,9 +6,11 @@
 package cn.hanbell.kpi.mail;
 
 import cn.hanbell.kpi.comm.MailNotification;
-import cn.hanbell.kpi.ejb.erp.GroupHSShipmentBean;
-import cn.hanbell.kpi.ejb.erp.GroupServiceBean;
-import cn.hanbell.kpi.ejb.erp.GroupShipmentBean;
+import cn.hanbell.kpi.ejb.erp.BscGroupHSSaleOrderBean;
+import cn.hanbell.kpi.ejb.erp.BscGroupHSShipmentBean;
+import cn.hanbell.kpi.ejb.erp.BscGroupSHSaleOrderBean;
+import cn.hanbell.kpi.ejb.erp.BscGroupSHServiceBean;
+import cn.hanbell.kpi.ejb.erp.BscGroupSHShipmentBean;
 import cn.hanbell.kpi.entity.Indicator;
 import java.util.Date;
 import java.util.List;
@@ -24,18 +26,18 @@ import tw.hanbell.kpi.ejb.erp.GrpsdailytmpBean;
  */
 @Stateless
 @LocalBean
-public class GroupShipmentMailBean extends MailNotification {
+public class BscGroupShipmentMailBean extends MailNotification {
 
     @EJB
-    private GroupShipmentBean groupShipmentBean;
+    private BscGroupSHShipmentBean bscGroupSHShipmentBean;
     @EJB
-    private GroupServiceBean groupServiceBean;
+    private BscGroupSHServiceBean bscGroupSHServiceBean;
     @EJB
-    private GroupHSShipmentBean groupHSShipmentBean;
+    private BscGroupHSShipmentBean bscGroupHSShipmentBean;
     @EJB
     private GrpsdailytmpBean grpsdailytmpBean;
 
-    public GroupShipmentMailBean() {
+    public BscGroupShipmentMailBean() {
 
     }
 
@@ -58,12 +60,15 @@ public class GroupShipmentMailBean extends MailNotification {
     @Override
     protected String getMailBody() {
         try {
-            //先更新汉钟的
-            groupShipmentBean.updataActualValue(y, m, d,"Shipment");
-            groupServiceBean.updataActualValue(y, m, d);
-            log4j.info("End Execute Job updateSHBBscGroupShipment");
-            groupHSShipmentBean.updataActualValue(y, m, d,"Shipment");
-            log4j.info("End Execute Job updateHansonBscGroupShipment");
+            //汉钟的出货
+            bscGroupSHShipmentBean.updataShpimentActualValue(y, m, d);
+            log4j.info("End Execute Job updataShpimentActualValue");
+            //汉钟的服务
+            bscGroupSHServiceBean.updataServerActualValue(y, m, d);
+            log4j.info("End Execute Job updataServerActualValue");
+            //汉声的出货
+            bscGroupHSShipmentBean.updataShpimentActualValue(y, m, d);
+            log4j.info("End Execute Job updataShpimentActualValue");
             //再更新到台湾ERP
             StringBuilder sb = new StringBuilder();
             int num;
@@ -80,7 +85,7 @@ public class GroupShipmentMailBean extends MailNotification {
             sb.append(getMailFooter());
             return "更新集团报表（汉钟中间表）数据成功" + sb.toString();
         } catch (Exception ex) {
-            return ex.toString() + "捕获异常，请检查排程！！！";    
+            return ex.toString() + "捕获异常，请检查排程！！！";
         }
     }
